refactor(deactivate): make exit an alias of deactivate
Consolidate the exit and deactivate commands by making exit an alias of deactivate.
