+++ cli/flox/src/commands/deactivate.rs
@@ -1,12 +1,12 @@
 use std::io::{BufWriter, stdout};
 
-use anyhow::{Result, bail};
+use anyhow::Result;
 use bpaf::Bpaf;
 use flox_rust_sdk::flox::Flox;
 use flox_rust_sdk::utils::FLOX_INTERPRETER;
-use indoc::indoc;
-use shell_gen::{Shell, ShellWithPath};
+use indoc::{formatdoc, indoc};
 
+use super::{activated_environments, uninitialized_environment_description};
 use crate::commands::activate::ActivateOptions;
 use crate::subcommand_metric;
 use crate::utils::message;
@@ -21,9 +21,7 @@ pub struct Deactivate {
 impl Deactivate {
     pub fn handle(self, flox: Flox) -> Result<()> {
         if !flox.features.auto_activate {
-            bail!(
-                "'flox deactivate' requires the auto_activate feature flag. Set FLOX_FEATURES_AUTO_ACTIVATE=true."
-            );
+            return self.old_exit(flox);
         }
 
         subcommand_metric!("deactivate");
@@ -44,14 +42,47 @@ impl Deactivate {
             Ok(())
         } else {
             // Interactive mode - print instructions
-            message::info(indoc! {"
-                To deactivate the current environment, type 'exit' to exit your shell.
+            let active_environments = activated_environments();
+            let last_active = active_environments.last_active();
 
-                Alternatively, you can restore environment variables with:
-                  eval \"$(flox deactivate --print-script)\"
-            "});
+            let Some(last_active) = last_active else {
+                message::info(indoc! {"
+                    No environment active!
+                    Exit active environments by typing 'exit' to exit your current shell or close your terminal.
+                    Environments can be activated using `flox activate`.
+                "});
+
+                return Ok(());
+            };
+
+            message::info(formatdoc! {"
+                Exit the currently active environment {} by typing 'exit' to exit your current shell or close your terminal.
+            ", uninitialized_environment_description(&last_active)?});
 
             Ok(())
         }
+    }
+
+    pub fn old_exit(self, _flox: Flox) -> Result<()> {
+        subcommand_metric!("exit");
+
+        let active_environments = activated_environments();
+        let last_active = active_environments.last_active();
+
+        let Some(last_active) = last_active else {
+            message::info(indoc! {"
+                No environment active!
+                Exit active environments by typing 'exit' to exit your current shell or close your terminal.
+                Environments can be activated using `flox activate`.
+            "});
+
+            return Ok(());
+        };
+
+        message::info(formatdoc! {"
+            Exit the currently active environment {} by typing 'exit' to exit your current shell or close your terminal.
+        ", uninitialized_environment_description(&last_active)?});
+
+        Ok(())
     }
 }
